Move AnnotationTestObject to test package
